[Changed index page to functional component and moved all business logic to redux reducer]
